Ported code to .netstandard
I ported all of our code to .netstandard1.0. This will allow for easy
transitions to new versions of .net. It can now run from 4.5 to Core. It
used to run only on 4.5.2 - 4.6. Although... I'm having trouble getting
it to work on core.

I needed to get a new assertion library because should isn't on
.netcore. So, I just created a bunch of mapping extensions to
FluentAssertions.